ohne ausgaben

+++ Computergraphik/Abgabe6/Blatt01/ModelHE.h
@@ -36,5 +36,5 @@ private:
 
 	std::vector<glm::vec3> vertices;
 	std::vector<glm::vec3> colors;
-	std::vector<GLushort> indices;
+	std::vector<GLuint> indices;
 };
